requirements.txt
tensorflow==2.10.*
pyautogui==0.9.52
pillow==8.4.0
tkcap==0.0.4
pydicom==2.2.2
img2pdf==0.4.4
opencv-python==4.5.5.64
matplotlib==3.5.1
pandas==1.3.5
python-xlib==0.31
protobuf==3.19.6
keras==2.10.*
pytest==7.2.2
